+++ fltk-sys/src/fl.rs
@@ -102,6 +102,9 @@ extern "C" {
     pub fn Fl_set_fonts(c: *const libc::c_char) -> libc::c_uchar;
 }
 extern "C" {
+    pub fn Fl_set_font(arg1: libc::c_int, arg2: libc::c_int);
+}
+extern "C" {
     pub fn Fl_add_handler(
         ev_handler: ::core::option::Option<unsafe extern "C" fn(ev: libc::c_int) -> libc::c_int>,
     );